move cmd/ to cli/
