+++ cli/plugins.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"github.com/vjeantet/bitfan/core"
